[Add] 사운드 파일 분리 중
